requirements.txt
aiogram==3.7.0
openpyxl==3.1.5
python-dotenv==1.0.1
playwright==1.53.0